jdk: Fix missing menu bar in AWT applications.
Turned out to be a regression of r333.

+++ trunk/openjdk/jdk/src/windows/native/sun/windows/awt_MenuBar.h
@@ -66,7 +66,7 @@ public:
 
     virtual HWND GetOwnerHWnd();
 
-    AwtMenuItem* GetItem(jobject target, long index);
+    AwtMenuItem* GetItem(jobject target, jint index);
     int CountItem(jobject menuBar);
 
     void SendDrawItem(AwtMenuItem* awtMenuItem,
